requirements.txt
#LangChain Core 
langchain 
langchain-core
langchain-community

#OpenAI 
langchain-openai
openai

#Gemini
langchain-google-genai
google-generativeai

#HugginFace
langchain-huggingface
transformers
huggingface_hub

#DotEnv
python-dotenv

#Machine Learning Utilities
numpy
scikit-learn

#Streamlit
streamlit

# Drawing Chains
grandalf

# Pdf 
pypdf
bs4
